feat(initia): refactor statesync and peers query logic
